+++ CHANGES.rst
@@ -1,5 +1,13 @@
 CHANGE LOG
 ==========
+10.1.3
+------
+Add API endpoint to display forms
+
+10.1.2
+------
+Fix skip logic's checkbox issue
+
 10.1.1
 ------
 Fix pagination on forms submission